refactor(resolver): dedupe forge phase-11 call-resolution helpers
Phase 11 decomposed build-edges.ts's resolveFallbackTargets into 15 named
helpers. The codebase-wide duplicate scan on those new helpers surfaced two
pre-existing duplicates the decomposition didn't catch (both predate phase 11
temporally, so the consolidation opportunities weren't nameable until now):

- resolveReflectionKeyExprFallback and resolveDefinePropertyAccessorFallback
  each inlined the same string/{type} typeMap-entry unwrap ternary that
  strategy.ts's unwrapTypeEntry already implements (extracted one commit
  after phase 11, in forge phase 20). Promoted unwrapTypeEntry to an export
  and wired both call sites to it.
- build-edges.ts's newly-named resolveSameClassQualifiedMethod was
  byte-identical to incremental.ts's pre-existing resolveThisSameClassTarget
  (incremental.ts's own docstring already noted it mirrors "the full-build
  counterpart"). Moved the shared implementation into call-resolver.ts --
  the module whose stated purpose is holding call-resolution logic shared
  between build-edges.ts and incremental.ts exactly once -- and wired both
  consumers to it.

Zero behavior change: all replaced logic was verified byte-identical before
relocating. Resolution benchmark (206/206) and full suite (3340 tests) pass
identically before/after.

Two follow-on divergences the scan also surfaced between the full-build and
incremental resolution paths (not fixed here -- would change behavior, not
just relocate it): incremental.ts has no same-class bare-call fallback
(#1765), and the Object.defineProperty same-file fallback's kind-filter
differs between the two paths (#1766).

docs check acknowledged: pure internal dedup/extract-method refactor, no
new features, commands, languages, or architecture changes -- README/CLAUDE/
ROADMAP do not need updates.

Impact: 5 functions changed, 20 affected
